+++ stickerfinder/config.py
@@ -30,7 +30,12 @@ default_config = {
         "port": 7000,
     },
     "job": {"user_check_count": 200, "report_count": 5,},
-    "mode": {"leecher": False, "authorized_only": False, "auto_accept_set": False,},
+    "mode": {
+        "leecher": False,
+        "authorized_only": False,
+        "auto_accept_set": False,
+        "private_inline_query": False,
+    },
 }
 
 config_path = os.path.expanduser("~/.config/stickerfinder.toml")
@@ -42,3 +47,9 @@ if not os.path.exists(config_path):
     sys.exit(1)
 else:
     config = toml.load(config_path)
+
+    # Set default values for any missing keys in the loaded config
+    for key, category in default_config.items():
+        for option, value in category.items():
+            if option not in config[key]:
+                config[key][option] = value